Development (#11)
* minor main page bugfixes

* value resolvers not working properly

* fixed it

* added value resolvers for getdocumentdto

* made share page work by using backend methods

* added updateDocument authorization, unit tests project and removed unused service

* added several DocumentService tests and fixes

- fixed value resolvers bugs
- fixed UpdateDocument and DeleteDocument authorization
- working on GetAllDocuments tests

* added more unit tests and ShareDocument service

- unit test for paging (PageResults)
- finished DocumentService unit tests
- moved document sharing logic to separate service
- working on tests for ShareDocument service

* added accountService

 - implemented username and password updating
- working on confirm email and reset password
- now deleting document removes all of its accessLevels

* fix conflicts

* Rebuild DocumentCard, Started impl DeleteDocument

* Finished implementing document delete (view)

* Moved tests to corresponding folder

* mail service - refactor needed

* mail service refactor

- added email confirmation and password reset in account service
- updated email messages
- added 'new password must be different from old one' policy

* added encryption of email link

- database tweaks
- added encryptionkey table
- submit email confirmation with encryption
- working on password reset encryption

* working password reset and email confirmation link encryption

* Added submit email confirmation on sign in page. Moved auth pages to according folders. Small encryption fixes. Need to implement event scheduler for encryption keys recreating.

* moved key creation to class EncryptionKeyGenerator.cs and made it a background service

* Added lockout disabling on email confirmation and boxed strings in SubmitEmailConfirmation.razor and SubmitPasswordReset.razor for optimal Radzen performance (no need for double click on submit button).

---------

+++ DocumentsApp.Data/Services/DocumentService.cs
@@ -6,6 +6,7 @@ using DocumentsApp.Data.Exceptions;
 using DocumentsApp.Data.Repos.Interfaces;
 using DocumentsApp.Data.Services.Interfaces;
 using DocumentsApp.Shared.Dtos.DocumentDtos;
+using DocumentsApp.Shared.Enums;
 using Microsoft.EntityFrameworkCore;
 using Sieve.Models;
 using Sieve.Services;
@@ -18,27 +19,27 @@ public class DocumentService : IDocumentService
     private readonly IDocumentRepo _documentRepo;
     private readonly IAccessLevelRepo _accessLevelRepo;
     private readonly ISieveProcessor _sieveProcessor;
-    private readonly IAuthenticationContextProvider _authenticationStateProvider;
+    private readonly IAuthenticationContextProvider _authenticationContextProvider;
 
     public DocumentService(
         IMapper mapper,
         IDocumentRepo documentRepo,
         ISieveProcessor sieveProcessor,
         IAccessLevelRepo accessLevelRepo,
-        IAuthenticationContextProvider authenticationStateProvider
+        IAuthenticationContextProvider authenticationContextProvider
     )
     {
         _mapper = mapper;
         _documentRepo = documentRepo;
         _accessLevelRepo = accessLevelRepo;
         _sieveProcessor = sieveProcessor;
-        _authenticationStateProvider = authenticationStateProvider;
+        _authenticationContextProvider = authenticationContextProvider;
     }
 
-    public async Task<GetDocumentDto> GetDocumentByIdAsync(string id)
+    public async Task<GetDocumentDto> GetDocumentByIdAsync(string documentId)
     {
-        var accountId = await _authenticationStateProvider.GetUserId();
-        var document = await SearchDocumentDbAsync(id);
+        var accountId = await _authenticationContextProvider.GetUserId();
+        var document = await SearchDocumentDbAsync(documentId);
         var documentAccessLevel = await _accessLevelRepo.GetDocumentAccessLevelAsync(accountId, document.Id);
 
         if (documentAccessLevel is null && document.AccountId != accountId)
@@ -50,12 +51,13 @@ public class DocumentService : IDocumentService
 
     public async Task<PagedResults<GetDocumentDto>> GetAllDocumentsAsync(SieveModel query)
     {
-        var accountId = await _authenticationStateProvider.GetUserId();
+        var accountId = await _authenticationContextProvider.GetUserId();
         var documents = _documentRepo.GetAllUserDocumentsAsQueryable(accountId);
         
-        if (!documents.Any()) throw new NotFoundException("No documents available for this user");
+        if (!documents.Any()) 
+            throw new NotFoundException("No documents available for this user");
 
-         var resultDocuments = await _sieveProcessor
+        var resultDocuments = await _sieveProcessor
             .Apply(query, documents)
             .Select(d => _mapper.Map<GetDocumentDto>(d))
             .ToListAsync();
@@ -64,52 +66,54 @@ public class DocumentService : IDocumentService
             resultDocuments,
             documents.Count(),
             query.PageSize.GetValueOrDefault(),
-            query.Page.GetValueOrDefault()
-            );
-    }
-
-    public async Task<PagedResults<GetDocumentDto>> GetAllSharedDocumentsAsync(SieveModel query)
-    {
-        var accountId = await _authenticationStateProvider.GetUserId();
-        var documents = _documentRepo.GetAllSharedDocumentsAsQueryable(accountId);
-        
-        if (!documents.Any()) throw new NotFoundException("No documents shared for this user");
-        
-        var resultDocuments = await _sieveProcessor
-            .Apply(query, documents)
-            .Select(d => _mapper.Map<GetDocumentDto>(d))
-            .ToListAsync();
-        
-        return new PagedResults<GetDocumentDto>(resultDocuments, documents.Count(),
-            query.PageSize.GetValueOrDefault(), query.Page.GetValueOrDefault());
+            query.Page.GetValueOrDefault());
     }
 
     public async Task<GetDocumentDto> AddDocumentAsync(AddDocumentDto dto)
     {
         var document = _mapper.Map<Document>(dto);
-        document.AccountId = await _authenticationStateProvider.GetUserId();
-        await _documentRepo.InsertDocumentAsync(document);
+        document.AccountId = await _authenticationContextProvider.GetUserId();
         
-        return _mapper.Map<GetDocumentDto>(document);
+        var result = await _documentRepo.InsertDocumentAsync(document);
+        
+        return _mapper.Map<GetDocumentDto>(result);
     }
 
-    public async Task UpdateDocumentAsync(string id, UpdateDocumentDto dto)
+    public async Task<GetDocumentDto> UpdateDocumentAsync(string documentId, UpdateDocumentDto dto)
     {
-        var document = await SearchDocumentDbAsync(id);
+        var document = await SearchDocumentDbAsync(documentId);
+        var userId = await _authenticationContextProvider.GetUserId();
+        var accessLevel = await _accessLevelRepo.GetDocumentAccessLevelAsync(userId, documentId);
+
+        if (document.AccountId != userId && accessLevel?.AccessLevelEnum != AccessLevelEnum.Write)
+            throw new NotAuthorizedException("User is not authorized to edit this document");
+
         _mapper.Map(dto, document);
-        await _documentRepo.UpdateDocumentAsync(document);
+
+        var result = await _documentRepo.UpdateDocumentAsync(document);
+        
+        return _mapper.Map<GetDocumentDto>(result);
     }
 
-    public async Task DeleteDocumentAsync(string id)
+    public async Task<bool> DeleteDocumentAsync(string documentId)
     {
-        var document = await SearchDocumentDbAsync(id);
-        await _documentRepo.DeleteDocumentAsync(document);
+        var document = await SearchDocumentDbAsync(documentId);
+        var userId = await _authenticationContextProvider.GetUserId();
+        
+        if (document.AccountId != userId)
+            throw new NotAuthorizedException("User is not authorized to delete this document");
+
+        await _accessLevelRepo.RemoveAllDocumentAccessLevelsAsync(documentId);
+        var result = await _documentRepo.DeleteDocumentAsync(document);
+        
+        return result;
     }
 
-    private async Task<Document> SearchDocumentDbAsync(string id)
+    private async Task<Document> SearchDocumentDbAsync(string documentId)
     {
-        var foundDocument = await _documentRepo.GetDocumentByIdAsync(id);
-        if (foundDocument is null) throw new NotFoundException("Document does not exist");
+        var foundDocument = await _documentRepo.GetDocumentByIdAsync(documentId);
+        if (foundDocument is null) 
+            throw new NotFoundException("Document does not exist");
 
         return foundDocument;
     }
